Replace FortAwesome components with CDN MD icons
Somehow, tree shaking wasn't working with FortAwesome components,
resulting in an over 500KB increase in my bundle size.
Hence, get rid of the FortAwesome components and switch to a CDN.
Might as well try out MD icons in the meantime.

Also, update Babel in an attempt to investigate the tree shaking issue.